requirements.txt
flask
flask-cors
streamlit
requests
feedparser
newspaper3k
python-dotenv

